+++ src/web-worker/WebWorkerPostMessageStream.ts
